requirements.txt
boto3
kydb
requests
PyNaCl